refactor: rootchain test only functions
