+++ apps/omg_eth/mix.exs
@@ -50,7 +50,9 @@ defmodule OMG.Eth.MixProject do
        git: "https://github.com/pthomalla/exexec.git", branch: "add_streams", only: [:dev, :test], runtime: false},
       {:briefly, "~> 0.3.0", only: [:dev, :test], runtime: false},
       {:websockex, "~> 0.4.2"},
-      {:omg_utils, in_umbrella: true}
+      {:omg_utils, in_umbrella: true},
+      # Used for mocking websocket servers
+      {:plug_cowboy, "~> 1.0", only: [:dev, :test]}
     ]
   end
