+++ apps/omg_watcher/test/support/integration/test_helper.ex
@@ -77,7 +77,7 @@ defmodule OMG.Watcher.Integration.TestHelper do
     Process.sleep(Application.fetch_env!(:omg_eth, :exit_period_seconds) * 1_000)
 
     {:ok, %{"blockNumber" => block_number}} =
-      OMG.Eth.RootChain.process_exits(token, 0, 1, user.addr) |> Eth.DevHelpers.transact_sync!()
+      OMG.Eth.RootChainHelper.process_exits(token, 0, 1, user.addr) |> Eth.DevHelpers.transact_sync!()
 
     Eth.DevHelpers.wait_for_root_chain_block(block_number + 2, 3_000)
   end
